drivers: Move client/drivers/executor to drivers/shared/executor
