+++ client/driver/executor_plugin.go
@@ -9,8 +9,8 @@ import (
 
 	hclog "github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/go-plugin"
-	"github.com/hashicorp/nomad/client/driver/executor"
 	cstructs "github.com/hashicorp/nomad/client/structs"
+	"github.com/hashicorp/nomad/drivers/shared/executor"
 )
 
 // Registering these types since we have to serialize and de-serialize the Task
